+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.2] - 2026-03-23
+
+### Changed
+
+- Pinned agent-client-protocol dependency back to 0.8.1
+
+### Removed
+
+- Context usage updates via ACP
+
+
 ## [2.6.1] - 2026-03-23
 
 ### Changed